requirements.txt
gTTS==2.3.2
pygame==2.5.0
pydub==0.25.1
google-generativeai